+++ public/components/contribution-month/contribution-month.js
@@ -2,7 +2,7 @@ import Component from 'can-component';
 import DefineMap from 'can-define/map/';
 import './contribution-month.less';
 import template from './contribution-month.stache';
-import ContributionMonth from 'bitcentive/models/contribution-month';
+import ContributionMonth from 'bitcentive/models/contribution-month/';
 
 export const ViewModel = DefineMap.extend({
   // Passed props
