+++ public/components/contributions/contributions.html
@@ -6,7 +6,7 @@
   <contributions {contribution-month}="contributionMonth"/>
 </script>
 <script src="../../../node_modules/steal/steal.js" main="can-view-autorender">
-  import ContributionMonth from 'bitcentive/models/contribution-month';
+  import ContributionMonth from 'bitcentive/models/contribution-month/';
   import OSProject from 'bitcentive/models/os-project';
   import osProjects from 'bitcentive/models/fixtures/contributions.json';
   import ClientProject from 'bitcentive/models/client-project';
