+++ public/components/contributions/contributions.js
@@ -3,7 +3,7 @@ import DefineMap from 'can-define/map/';
 import './contributions.less';
 import template from './contributions.stache';
 import MonthlyContribution from '../../models/monthly-contributions';
-import ContributionMonth from '../../models/contribution-month';
+import ContributionMonth from '../../models/contribution-month/';
 import Contributor from '../../models/contributor';
 import OSProject from '../../models/os-project';
