+++ public/components/os-projects/os-projects.html
@@ -7,7 +7,7 @@
 </script>
 <script src="../../../node_modules/steal/steal.js" main="can-view-autorender">
   import 'bitcentive/models/fixtures/fixtures-socket';
-  import ContributionMonth from 'bitcentive/models/contribution-month';
+  import ContributionMonth from 'bitcentive/models/contribution-month/';
   import viewModel from 'can-view-model';
 
   ContributionMonth.get("1").then(function(item){
