+++ public/components/os-projects/os-projects.js
@@ -3,7 +3,7 @@ import DefineMap from 'can-define/map/';
 import './os-projects.less';
 import template from './os-projects.stache';
 import OSProject from '../../models/os-project';
-import ContributionMonth from '../../models/contribution-month';
+import ContributionMonth from '../../models/contribution-month/';
 
 export const ViewModel = DefineMap.extend({
   // Passed properties
