+++ public/components/select-contribution-month/select-contribution-month.js
@@ -3,7 +3,7 @@ import Component from 'can-component';
 import DefineMap from 'can-define/map/';
 import './select-contribution-month.less';
 import template from './select-contribution-month.stache';
-import ContributionMonth from 'bitcentive/models/contribution-month';
+import ContributionMonth from 'bitcentive/models/contribution-month/';
 
 export const ViewModel = DefineMap.extend({
   selectedContributionMonthId: {
