+++ public/models/fixtures/fixtures.js
@@ -1,5 +1,5 @@
 import fixture from "can-fixture";
-import ContributionMonth from 'bitcentive/models/contribution-month';
+import ContributionMonth from 'bitcentive/models/contribution-month/';
 import OSProject from 'bitcentive/models/os-project';
 import ClientProject from 'bitcentive/models/client-project';
